refactor: DuckTestSmallCover를 shared 모듈로 이동
